@@ -2,7 +2,7 @@
 name = "stonemq"
 version = "0.1.0"
 edition = "2021"
-rust-version = "1.73.0"
+rust-version = "1.80.0"
 repository = "https://github.com/jonefeewang/stonemq"
 authors = ["jonefeewang<jonefeewang@gmail.com>"]
 description = "A high performance and efficient message queue written in Rust."
@@ -41,7 +41,6 @@ uuid = { version = "1.1.2", features = ["v4"] }
 rand = "0.8.5"
 integer-encoding = "4.0.0"
 crc32c = "0.6.5"
-once_cell = "1.19.0"
 memmap2 = "0.9.4"
 dotenv = "0.15.0"
 
