@@ -10,11 +10,10 @@ use crate::{
     AppError,
 };
 use std::collections::HashMap;
-use std::sync::Arc;
+use std::sync::{Arc, LazyLock};
 
 use crate::{message::TopicPartition, AppResult};
 use bytes::{BufMut, BytesMut};
-use once_cell::sync::Lazy;
 
 const GROUP_ID_KEY_NAME: &str = "group_id";
 
@@ -210,13 +209,13 @@ impl FetchOffsetsResponse {
     }
 }
 
-pub static OFFSET_FETCH_REQUEST_PARTITION_V0_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
+pub static OFFSET_FETCH_REQUEST_PARTITION_V0_SCHEMA: LazyLock<Arc<Schema>> = LazyLock::new(|| {
     let fields_desc: Vec<(i32, &str, ProtocolType)> =
         vec![(0, PARTITION_KEY_NAME, ProtocolType::I32(I32::default()))];
     Arc::new(Schema::from_fields_desc_vec(fields_desc))
 });
 
-pub static OFFSET_FETCH_REQUEST_TOPIC_V0_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
+pub static OFFSET_FETCH_REQUEST_TOPIC_V0_SCHEMA: LazyLock<Arc<Schema>> = LazyLock::new(|| {
     let fields_desc: Vec<(i32, &str, ProtocolType)> = vec![
         (0, TOPIC_KEY_NAME, ProtocolType::PString(PString::default())),
         (
@@ -235,7 +234,7 @@ pub static OFFSET_FETCH_REQUEST_TOPIC_V0_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|
     Arc::new(Schema::from_fields_desc_vec(fields_desc))
 });
 
-pub static OFFSET_FETCH_REQUEST_V3_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
+pub static OFFSET_FETCH_REQUEST_V3_SCHEMA: LazyLock<Arc<Schema>> = LazyLock::new(|| {
     let fields_desc: Vec<(i32, &str, ProtocolType)> = vec![
         (
             0,
@@ -257,7 +256,7 @@ pub static OFFSET_FETCH_REQUEST_V3_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
     Arc::new(Schema::from_fields_desc_vec(fields_desc))
 });
 
-pub static OFFSET_FETCH_RESPONSE_PARTITION_V0_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
+pub static OFFSET_FETCH_RESPONSE_PARTITION_V0_SCHEMA: LazyLock<Arc<Schema>> = LazyLock::new(|| {
     let fields_desc: Vec<(i32, &str, ProtocolType)> = vec![
         (0, PARTITION_KEY_NAME, ProtocolType::I32(I32::default())),
         (1, OFFSET_KEY_NAME, ProtocolType::I64(I64::default())),
@@ -271,7 +270,7 @@ pub static OFFSET_FETCH_RESPONSE_PARTITION_V0_SCHEMA: Lazy<Arc<Schema>> = Lazy::
     Arc::new(Schema::from_fields_desc_vec(fields_desc))
 });
 
-pub static OFFSET_FETCH_RESPONSE_TOPIC_V0_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
+pub static OFFSET_FETCH_RESPONSE_TOPIC_V0_SCHEMA: LazyLock<Arc<Schema>> = LazyLock::new(|| {
     let fields_desc: Vec<(i32, &str, ProtocolType)> = vec![
         (0, TOPIC_KEY_NAME, ProtocolType::PString(PString::default())),
         (
@@ -289,7 +288,7 @@ pub static OFFSET_FETCH_RESPONSE_TOPIC_V0_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(
     Arc::new(Schema::from_fields_desc_vec(fields_desc))
 });
 
-pub static OFFSET_FETCH_RESPONSE_V3_SCHEMA: Lazy<Arc<Schema>> = Lazy::new(|| {
+pub static OFFSET_FETCH_RESPONSE_V3_SCHEMA: LazyLock<Arc<Schema>> = LazyLock::new(|| {
     let fields_desc: Vec<(i32, &str, ProtocolType)> = vec![
         (0, THROTTLE_TIME_KEY_NAME, ProtocolType::I32(I32::default())),
         (
